# openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Puyuan AI Platform API (MVP)
  version: 1.0.0
  description: MVP API for merchant platform, billing, plugins and admin operations.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
...
paths:
...
  /api/v1/plugins:
    get:
      tags: [Plugins]
      summary: Query available plugins
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Plugin list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PluginItem'
  /api/v1/plugins/{plugin_id}/enable:
    post:
      tags: [Plugins]
      summary: Enable plugin for tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Enable result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/plugins/{plugin_id}/disable:
    post:
      tags: [Plugins]
      summary: Disable plugin for tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Disable result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/plugins/{plugin_id}/invoke:
    post:
      tags: [Plugins]
      summary: Invoke plugin
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginInvokeRequest'
      responses:
        '200':
          description: Invocation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PluginInvokeResponse'
...
  /api/v1/admin/plugins/{plugin_id}:
    patch:
      tags: [Admin]
      summary: Update plugin
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: plugin_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePluginRequest'
      responses:
        '200':
          description: Plugin updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenantId:
      in: header
      name: X-Tenant-Id
      required: true
      schema:
        type: string
    XRequestId:
      in: header
      name: X-Request-Id
      required: true
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          nullable: true
        request_id:
          type: string
      required: [code, message, request_id]
...
    PluginItem:
      type: object
      properties:
        plugin_id:
          type: string
        name:
          type: string
        version:
          type: string
        billing_type:
          type: string
          enum: [token, times, free]
        enabled:
          type: boolean
    PluginInvokeRequest:
      type: object
      description: Plugin custom payload.
      additionalProperties: true
    PluginInvokeResponse:
      type: object
      properties:
        result:
          type: object
          additionalProperties: true
        token_used:
          type: integer
        balance_remaining:
          type: integer
...
    UpdatePluginRequest:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        backend_api:
          type: string
        frontend_entry:
          type: string
        billing_type:
          type: string
          enum: [token, times, free]
...